Added a nop function mb_lmm_init
This function will eventually do the initialization of the lmm
(list memory manager), which is the kernel's dynamic memory
manager. We need to tell lmm what regions of memory it is allowed
to use, and we use the information given to us in the multiboot_info
struct to do this.