Sprinkled debug/assert code throughout kernel (still crashing...)
I have at least simplified how to reproduce the bug. It seems to
occur when you are printing and getting interrupted while printing.
I don't know if it needs to happen many times, or just once at the
"right" (wrong) time, or what. It does not matter if you are
printing to the VGA console or to the debug console. The crash
only occurs in BOCHS. It does not crash in QEMU...