Put gdt/idt/tss into own section to prevent future bugs (read on)
Putting these data structures in the .text section was causing them to
be shuffled around. They are now in their own .intel section which
will start at 0x10000c. I added asserts to mb_entry() to make sure
we fail if they move around at all.

We are now able to get to userspace but we page fault trying to read
the first instruction at 0x40000000.